backend/requirements.txt
pandas>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
argon2-cffi>=23.1.0
itsdangerous>=2.2.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
alembic>=1.13.1
numpy>=1.24
sentence-transformers>=3.0.0
SQLAlchemy>=2.0
